requirements.txt
agno
lancedb
openai
python-dotenv
qmd
ag-ui-protocol
uvicorn
fastapi
